refactor(captioner): split captioner.py into package for lazy loading heavy deps
